Resolvendo erro de andressa

+++ src/main/java/br/com/fujideia/iesp/tecback/repository/SerieRepository.java
@@ -1,9 +1,10 @@
 package br.com.fujideia.iesp.tecback.repository;
 
-import br.com.fujideia.sim.tecback.modelo.Serie;
+
+import br.com.fujideia.iesp.tecback.model.Serie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SerieRepository extends JpaRepository<Serie,Long>{
+public interface SerieRepository extends JpaRepository<Serie, Integer>{
         }
